+++ integration-tests/src/test/java/org/apache/druid/tests/indexer/AbstractITBatchIndexTest.java
@@ -467,8 +467,8 @@ public abstract class AbstractITBatchIndexTest extends AbstractIndexerTest
         "Segment count"
     );
     ITRetryUtil.retryUntilEquals(
-        () -> coordinator.getAvailableSegments(dataSource + config.getExtraDatasourceNameSuffix())
-                         .stream().filter(DataSegment::isTombstone).count(),
+        () -> (int) coordinator.getAvailableSegments(dataSource + config.getExtraDatasourceNameSuffix())
+                               .stream().filter(DataSegment::isTombstone).count(),
         numExpectedTombstones,
         "Tombstone count"
     );
